Add image scaler. Part of display monster code is there

+++ src/main/java/hellofx/fheroes2/kingdom/World.java
@@ -35,7 +35,7 @@ public class World {
         Instance = new World();
     }
 
-    public short w, h;
+    public int w, h;
 
     private World() {
     }
